Move db from pkg to internal
