+++ internal/db/abi/test_assets/ABI-1.json
