+++ internal/db/abi/test_assets/ABI-2.json
