+++ internal/db/abi/test_assets/ABI-3.json
