+++ internal/db/abi/test_assets/ABI-4.json
