+++ internal/db/state/test_assets/contract-codes/Code-2.json
