+++ internal/db/state/test_assets/contract-codes/Code-3.json
