+++ internal/db/state/test_assets/contract-codes/Code-4.json
