+++ internal/db/transaction/manager_test.go
@@ -1,7 +1,7 @@
 package transaction
 
 import (
-	"github.com/NethermindEth/juno/pkg/db"
+	"github.com/NethermindEth/juno/internal/db"
 	"math/big"
 	"testing"
 )
